[BE] refactor: Playinghistoryredisbuffer 이름 변경  #763
