+++ backend/app/src/test/java/com/onair/hearit/app/playinghistory/infrastructure/buffer/PlayingHistoryBufferFacadeTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.onair.hearit.app.fixture.DbHelper;
 import com.onair.hearit.app.playinghistory.infrastructure.buffer.converter.PlayingHistoryConverter;
-import com.onair.hearit.app.playinghistory.infrastructure.buffer.storage.RedisPlayingHistoryStorage;
 import com.onair.hearit.core.config.DataSourceConfig;
 import com.onair.hearit.core.domain.Category;
 import com.onair.hearit.core.domain.Hearit;
@@ -15,8 +14,6 @@ import com.onair.hearit.core.fixture.TestJpaAuditingConfig;
 import com.onair.hearit.core.infrastructure.jdbc.PlayingHistoryCommandRepository;
 import com.onair.hearit.core.infrastructure.jpa.HearitRepository;
 import com.onair.hearit.core.infrastructure.jpa.PlayingHistoryRepository;
-import java.lang.reflect.Field;
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
